todo ok

+++ Application/Models/SaleRequest.cs
@@ -9,6 +9,6 @@ namespace Application.Models
     public class SaleRequest
     {
         public List<SaleProductRequest> Products { get; set; }
-        public double TotalPayed { get; set; }
+        public decimal TotalPayed { get; set; }
     }
 }
